rust: Drop unnecessary casts
New stable Clippy points out the following:

     error: casting to the same type is unnecessary (`usize` -> `usize`)
       --> src/wrappers/themis/rust/src/keygen.rs:108:29
        |
    108 |         private_key.set_len(private_key_len as usize);
        |                             ^^^^^^^^^^^^^^^^^^^^^^^^ help: try: `private_key_len`
        |
        = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#unnecessary_cast
        = note: `-D clippy::unnecessary-cast` implied by `-D warnings`

+++ src/wrappers/themis/rust/src/secure_message.rs
@@ -183,7 +183,7 @@ impl SecureMessage {
                 return Err(error);
             }
             debug_assert!(encrypted_len <= encrypted.capacity());
-            encrypted.set_len(encrypted_len as usize);
+            encrypted.set_len(encrypted_len);
         }
 
         Ok(encrypted)
@@ -233,7 +233,7 @@ impl SecureMessage {
                 return Err(error);
             }
             debug_assert!(decrypted_len <= decrypted.capacity());
-            decrypted.set_len(decrypted_len as usize);
+            decrypted.set_len(decrypted_len);
         }
 
         Ok(decrypted)
@@ -365,7 +365,7 @@ impl SecureSign {
                 return Err(error);
             }
             debug_assert!(signed_len <= signed.capacity());
-            signed.set_len(signed_len as usize);
+            signed.set_len(signed_len);
         }
 
         Ok(signed)
@@ -493,7 +493,7 @@ impl SecureVerify {
                 return Err(error);
             }
             debug_assert!(original_len <= original.capacity());
-            original.set_len(original_len as usize);
+            original.set_len(original_len);
         }
 
         Ok(original)
